Setting to let user silence +python error.
If the user kicks off a version of vim with the jedi-vim plugin without
having Python support compiled in, it gives an error which halts vim
from starting up until it is dismissed. This commit adds a way to opt
out of that warning for people who use the same configuration and
plugins on different machines and different versions of vim.